Print out Dashboard contents

This function prints out the Dashboard contents of the target workspace. You can provide either workspaceName or analysis. If both values are provided, this function will use workspaceName argument over analysis argument.

getDashboard(workspaceName = "", analysis = NULL)

Arguments

workspaceName

The name of the workspace you want to get the overview provided through the Dashboard.

analysis

The name of the analysis use want to check the Dashboard of. The list of available analyses can be found with availableAnalysis().

Value

The last modified date as a message, followed by the Dashboard contents from the target workspace.

Examples

library(AnVIL)
if (gcloud_exists() && nzchar(avworkspace_name())) {
getDashboard(analysis = "salmon")
getDashboard(workspaceName = "Bioconductor-Workflow-DESeq2")
}