List outputs

List outputs

listOutput(
  workspaceName,
  submissionId = NULL,
  keyword = NULL,
  includeMetadata = FALSE,
  accountEmail = gcloud_account(),
  billingProjectName = gcloud_project()
)

Arguments

workspaceName

Name of the workspace

submissionId

Submission Id. If it's not provided, the most recent submission id with the 'succeeded' status will be used.

keyword

A character string containing a regular expression to be matched in the output file name. Under the default NULL, all the outputs from the workflow, including log files, will be returned.

includeMetadata

Under the default (FALSE), metadata files (e.g. stderr, stdout, .log, .sh), will not be returned.

accountEmail

Email linked to Terra account

billingProjectName

Name of the billing project

Value

A tibble with four columns

file : character() 'base name' of the file in the bucket.

workflow : character() name of the workflow the file is associated with.

task : character() name of the task in the workflow that generated the file.

path : character() full path to the file in the google bucket.