Launch Terra workflow

Launch Terra workflow

runWorkflow(
  workspaceName,
  workflowName = NULL,
  useCallCache = TRUE,
  inputName = NULL
)

Arguments

workspaceName

Name of the workspace that contains the workflow(s) you want to launch.

workflowName

Name of the workflow to run.

useCallCache

A logical. Under the default condition (TRUE), call cache will be used.

inputName

Name of you input entity. If the workflow is using Terra's data model, this is required. The available entities can be found using the findInputName function.

Value

This function will print out whether the call for workflow launching was successful or not.

Examples

library(AnVIL)
if (gcloud_exists() && nzchar(avworkspace_name())) {
if ("salmon" %in% avworkspaces()$name)
runWorkflow(workspaceName = "salmon")
}